microservices/app/src/requirements.txt
flask
requests
gunicorn

# add new packages one per each line

wit
python_dotenv
bs4
